Move parsing methods to parser.go
